Fix error on osquery configuration parsing due to signed integer over…

+++ src/shared/file_op.c
@@ -3293,7 +3293,7 @@ char * abspath(const char * path, char * buffer, size_t size) {
 }
 
 /* Return the content of a file from a given path */
-char * w_get_file_content(const char * path, long max_size) {
+char * w_get_file_content(const char * path, unsigned long max_size) {
     FILE * fp = NULL;
     char * buffer = NULL;
     long size;
@@ -3318,7 +3318,7 @@ char * w_get_file_content(const char * path, long max_size) {
     }
 
     // Check file size limit
-    if (size > max_size) {
+    if ((unsigned long)size > max_size) {
         mdebug1("Cannot load file '%s': it exceeds %ld MiB", path, (max_size / (1024 * 1024)));
         goto end;
     }
